Move ProjectDataBuilder logic to Projects datasource
